feat: integrate new table provider with DataSink
